Fixed a crash case when there is no account.
I had a crash occur on the app.ens.domains site when attempting to use a DNS record as an ENS name. I'm not quite sure where the actual error is, but this surfaced as `TypeError: Cannot read properties of undefined (reading 'toLowerCase')` at this line, it seems `account` can sometimes be undefined.

This error presented itself as blanking the entire screen, so not a great user experience. Any attempts to use the site for my DNS record experiences this error, resulting in a blank screen. So this doesn't really solve the problem, but this should hopefully prevent the entire site from crashing to a blank screen.

I'm guessing my DNS entry is misconfigured, as mentioned in https://github.com/ensdomains/ens-app/issues/1284 .