Allow previous registrants to reregister expired domains
closes https://github.com/ensdomains/ens-app/issues/928

Currently, an expired domain cannot be reregistered by the same owner.
This limitation is because the `isDeedOwner` check makes the
registration availability come back false for the previous registrant.

I'd love to be able to reregister domains I lost track of ;)

This change removes the isDeedOwner check from this logic. There may
unintended consequences of this chane and I am not sure
how best to test this change. I am very open to feedback. Thanks!